+++ src/component/provisioning/vendor.js
@@ -1,6 +1,7 @@
 import React  from 'react';
 
 import Layout from '../Layout';
+import './provisioning.scss';
 const slugify = require('../../builder/slugify');
 
 const breadcrumbs = [{
