+++ src/component/provisioning/vendors.js
@@ -1,54 +1,23 @@
-import React, { useState } from 'react';
+import React from 'react';
 
 import Layout from '../Layout';
-import { buildTable } from './phone';
-const slugify = require('../../builder/slugify');
+import TableOfContents from '../uc-doc/TableOfContents';
+import Vendors from './components/Vendors';
+import './provisioning.scss';
 
-const title = 'Provd Plugins';
+const breadcrumbs = [{ link: '/uc-doc/ecosystem/supported_devices', label: 'Supported Devices', active: true }]
+export default ({ pageContext: { plugins, images, content, title = 'Supported Devices' }}) => (
+  <Layout pageTitle="Supported Devices" PageTitleComponent="p" breadcrumbs={breadcrumbs} className="template-uc-doc contribute">
+    <div className="container principal">
+      <div className="main-content main-content-on-right-side">
+        <div className="main-content-left-col">
+          <TableOfContents />
+        </div>
 
-export default ({ pageContext: { plugins, images } }) => {
-  const [currentVendor, setCurrentVendor] = useState(null);
-  const [currentDevice, setCurrentDevice] = useState(null);
-
-  const onDeviceClick = event => {
-    event.preventDefault();
-    setCurrentDevice(currentDevice === event.target.id ? null : event.target.id);
-  }
-
-  const onVendorClick = event => {
-    event.preventDefault();
-    setCurrentVendor(currentVendor === event.target.id ? null : event.target.id);
-  }
-
-  return (
-    <Layout pageTitle={title}>
-      <div className="doc-wrapper provisioning-vendors">
-        <div className="container">
-          <div className="section-block">
-            {Object.keys(plugins).map(vendor => (
-              <div className="vendor" key={vendor}>
-                <a className="title" href={`/provisioning/${slugify(vendor)}`} id={vendor} onClick={onVendorClick}>{vendor}</a>
-                {vendor === currentVendor && 
-                  <div className="phones">
-                    {Object.keys(plugins[vendor]).map(name => {
-                      const id = `${slugify(vendor)}-${slugify(name)}`;
-                      return <div className={id === currentDevice ? 'selected' : ''} key={id} >
-                        <a href={`/provisioning/${slugify(vendor)}/${slugify(name)}`} onClick={onDeviceClick} id={id}>{name}</a>
-                        {id === currentDevice && <div className="slide">
-                          <div className="image">
-                            <div className="name">{name}</div>
-                            {images[slugify(vendor)].indexOf(`${slugify(name)}.png`) !== -1 ? <img src={`/provisioning/${slugify(vendor)}-${slugify(name)}.png`} alt={`${slugify(vendor)}-${name}`}/> : <img src='/provisioning/img-placeholder.png' alt={`${slugify(vendor)}-${name}`} />}
-                          </div>
-                          <div className="content">{buildTable(plugins[vendor][name])}</div>
-                    </div>}
-                      </div>;
-                    })}
-                </div>}
-              </div>
-            ))}
-          </div>
+        <div className="provisioning-vendors platform">
+          <Vendors plugins={plugins} images={images} />
         </div>
       </div>
-    </Layout>
-  );
-};
+    </div>
+  </Layout>
+);
